+++ MiniApp/Classes/core/Display/MiniAppWebView.swift
@@ -12,19 +12,42 @@ internal class MiniAppWebView: WKWebView {
         return config
     }
 
-    convenience init(miniAppId: String, versionId: String) {
+    convenience init(miniAppId: String, versionId: String, queryParams: String? = nil) {
         let schemeName = Constants.miniAppSchemePrefix + miniAppId
-        let urlRequest = URLRequest(url: URL(string: schemeName + "://miniapp/" + Constants.rootFileName)!)
         let config = MiniAppWebView.defaultConfig()
         config.setURLSchemeHandler(URLSchemeHandler(versionId: versionId), forURLScheme: schemeName)
         self.init(frame: .zero, configuration: config)
-        commonInit(urlRequest: urlRequest)
+        commonInit(urlRequest: Self.getURLRequest(miniAppId: miniAppId, schemeName: schemeName, queryParams: queryParams))
     }
 
     convenience init(miniAppURL: URL) {
         let urlRequest = URLRequest(url: miniAppURL)
         self.init(frame: .zero, configuration: MiniAppWebView.defaultConfig())
         commonInit(urlRequest: urlRequest)
+    }
+
+    private static func getURL(miniAppURL: URL, queryParams: String?) -> URL {
+        guard let urlWithQueryParam = miniAppURL.appendingPathComponent(Self.getQueryParams(queryParams: queryParams)).absoluteString.removingPercentEncoding else {
+            return miniAppURL
+        }
+
+        return URL(string: urlWithQueryParam)!
+    }
+
+    private static func getURLRequest(miniAppId: String, schemeName: String, queryParams: String?) -> URLRequest {
+        let baseUrl = schemeName + "://miniapp/" + Constants.rootFileName
+        guard let url = URL(string: baseUrl + MiniAppWebView.getQueryParams(queryParams: queryParams)) else {
+            MiniAppLogger.e("MiniAppWebView failed to parse Query Parameters, trying to load using only URL")
+            return URLRequest(url: URL(string: schemeName + "://miniapp/" + Constants.rootFileName)!)
+        }
+        return URLRequest(url: url)
+    }
+
+    private static func getQueryParams(queryParams: String?) -> String {
+        guard let paramsString = queryParams, !paramsString.isEmpty else {
+            return ""
+        }
+        return "?" + paramsString.replacingOccurrences(of: " ", with: "%20")
     }
 
     private func commonInit(urlRequest: URLRequest) {