+++ CHANGELOG.md
@@ -5,9 +5,14 @@
 **SDK**
 
 - **Feature** Added analytics for Mini App usage tracking
+- **Feature** Updated `getUserName` and `getProfilePhoto` interfaces to be asynchronous. Old methods are deprecated
+- **Feature** Updated `MiniApp().shared.create` interface to accept another optional query string parameter.
+- **Deprecated:** `requestPermission(permissionType:completionHandler:)` in `MiniAppMessageDelegate` protocol is deprecated. You should use `requestDevicePermission(permissionType:completionHandler:))` instead
 
 **Sample App**
 
+- **Change** Updated sample app to use the latest asynchronous `getUserName` and `getProfilePhoto` interfaces
+- **Feature:** Added a new section in Settings page to add the optional query parameter.
 
 ---